requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pillow==10.3.0
pytesseract==0.3.10
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.2.1
starlette==0.37.2
opencv-python-headless==4.9.0.80
psycopg2-binary==2.9.9
